Revert "Enhance PlotNado with new features, examples, and improvements (#33)"
This reverts commit 5f934ad3700792ddd3c06eea02ececb10c50662c.
