+++ _quarto.yml
@@ -29,8 +29,6 @@ website:
             href: docs/quickstart.qmd
           - text: Track Construction
             href: docs/quickstart_tracks.qmd
-      - text: Examples
-        href: docs/examples.qmd
       - text: Guides
         menu:
           - text: Track Catalog
@@ -75,7 +73,7 @@ format:
     toc-depth: 4
     code-copy: true
     code-overflow: wrap
-    page-layout: article
+    page-layout: full
 
 execute:
   enabled: true
