+++ docs/quickstart_tracks.qmd
@@ -12,30 +12,27 @@ Use helper methods for most figures, `add_track()` when names come from configur
 ## Helper methods
 
 ```{python}
-#| fig-cap: "Helper methods keep ordinary figure construction readable and chainable."
 from plotnado import GenomicFigure
-from plotnado.examples import REGION, intervals, signal
-# signal() → DataFrame(chrom, start, end, value) — replace with a BigWig path/URL or DataFrame
-# intervals() → DataFrame(chrom, start, end, name) — replace with a BED/BigBed path, URL, or DataFrame
+from docs._examples import REGION, intervals, signal
 
 fig = GenomicFigure(track_height=1.2)
 fig.autocolor("Set2")
 fig.scalebar()
+fig.axis()
 fig.bigwig(signal(), title="Signal", style="fill", color_group="sample-a")
 fig.bed(intervals(), title="Intervals", display="expanded", show_labels=True, color_group="sample-a")
-fig.axis()
 fig.plot(REGION)
 ```
+
+Caption: helper methods keep ordinary figure construction readable and chainable.
 
 ## Alias entry point
 
 `add_track()` is useful when track names are read from YAML, TOML, or another runtime source.
 
 ```{python}
-#| fig-cap: "Aliases map to the same track constructors used by helper methods."
 from plotnado import GenomicFigure
-from plotnado.examples import REGION, signal
-# signal() → DataFrame(chrom, start, end, value) — replace with a BigWig path/URL or DataFrame
+from docs._examples import REGION, signal
 
 fig = GenomicFigure().autocolor("Dark2")
 fig.add_track("scalebar")
@@ -45,14 +42,15 @@ fig.add_track("bigwig", data=signal(phase=0.8), title="Replicate B", alpha=0.75)
 fig.plot(REGION)
 ```
 
+Caption: aliases map to the same track constructors used by helper methods.
+
 ## Explicit objects
 
 Use explicit classes when you need to pass track objects around before adding them.
 
 ```python
 from plotnado import BigWigTrack, GenomicFigure
-from plotnado.examples import signal
-# signal() → DataFrame(chrom, start, end, value) — replace with a BigWig path/URL or DataFrame
+from docs._examples import signal
 
 track = BigWigTrack(data=signal(), title="Reusable signal", style="fill")
 fig = GenomicFigure().add_track(track)
@@ -88,26 +86,13 @@ fig.bigwig(
 Treat the overlay as the panel that owns the shared scale.
 
 ```{python}
-#| fig-cap: "Ordinary signal tracks and the overlay share one `autoscale_group`."
-from plotnado import GenomicFigure
-from plotnado.examples import REGION, review_signal
-# review_signal() → DataFrame(chrom, start, end, value) — replace with BigWig paths/URLs or DataFrames
+from docs._examples import REGION, overlay_comparison
 
-fig = GenomicFigure(track_height=1.2)
-fig.autoscale(True)
-fig.highlight("chr1:1,032,000-1,046,000")
-fig.highlight_style(color="#ffdd57", alpha=0.22)
-fig.bigwig(review_signal(2.0), title="Control", autoscale_group="signal", color="#1f77b4")
-fig.bigwig(review_signal(10.0, 1.2), title="Treatment", autoscale_group="signal", color="#d62728")
-fig.overlay(
-    [review_signal(5.5, 2.0), review_signal(6.5, 2.8)],
-    title="Overlay",
-    autoscale_group="signal",
-    colors=["#2ca02c", "#9467bd"],
-    alpha=0.55,
-)
+fig = overlay_comparison()
 fig.plot(REGION)
 ```
+
+Caption: ordinary signal tracks and the overlay share one `autoscale_group`.
 
 ## Option lookup
