+++ docs/track_aliases.qmd
@@ -12,10 +12,8 @@ execute:
 ## Alias usage
 
 ```{python}
-#| fig-cap: "Alias-added tracks render the same way as helper-method tracks."
 from plotnado import GenomicFigure
-from plotnado.examples import REGION, signal
-# signal() → DataFrame(chrom, start, end, value) — replace with a BigWig path/URL or DataFrame
+from docs._examples import REGION, signal
 
 fig = GenomicFigure().autocolor("Set2")
 fig.add_track("scalebar")
@@ -24,6 +22,8 @@ fig.add_track("bigwig", data=signal(phase=0.0), title="Replicate A", style="fill
 fig.add_track("bigwig", data=signal(phase=0.9), title="Replicate B", style="fill")
 fig.plot(REGION)
 ```
+
+Caption: alias-added tracks render the same way as helper-method tracks.
 
 ## How kwargs are routed
