+++ plotnado/__init__.py
@@ -6,7 +6,6 @@ without heavy dependencies like CoolBox.
 """
 
 from .figure import GenomicFigure
-from .tracks.genes import register_genome
 from .igv import parse_igv_session, IgvSession
 from .hub import parse_ucsc_hub, UcscHubSession
 from .widgets import TrackVisibilityWidget
@@ -15,7 +14,6 @@ from .template import Template, TrackSpec, GuideSpec, GroupSpec
 from .render import TemplateCompiler, RenderPlan, ResolvedTrack
 from .tracks import (
     # Enums
-    BedLabelPosition,
     CollectionStyle,
     DisplayMode,
     FontWeight,
@@ -83,7 +81,6 @@ except ImportError:
 
 __all__ = [
     "GenomicFigure",
-    "register_genome",
     "parse_igv_session",
     "IgvSession",
     "parse_ucsc_hub",
@@ -100,7 +97,6 @@ __all__ = [
     "RenderPlan",
     "ResolvedTrack",
     # Enums
-    "BedLabelPosition",
     "DisplayMode",
     "CollectionStyle",
     "FontWeight",
