--- plotnado/data/blueprint_monocyte_lyz_candidate_peaks.bed
@@ -1,4 +1,0 @@
-chr12	69344737	69344992	LYZ_promoter_peak
-chr12	69347375	69347611	LYZ_tss_peak
-chr12	69352454	69352660	LYZ_gene_body_peak
-chr12	69359719	69360538	LYZ_downstream_peak
