+++ plotnado/figure.pyi
@@ -114,8 +114,6 @@ class GenomicFigure:
             smoothing_center: bool = True,
             min_value: float | None = None,
             max_value: float | None = None,
-            n_bins: int | None = None,
-            bin_size: int | None = None,
             plot_title: bool = True,
             plot_scale: bool = True,
             label_on_track: bool = False,
@@ -208,7 +206,7 @@ class GenomicFigure:
             autoscale_group: str | None = None,
             color_group: str | None = None,
             show_chromosome: bool = False,
-            color: str = 'black',
+            color: str = 'steelblue',
             alpha: float = 1.0,
             linewidth: float = 1.0,
             font_size: int = 9,
@@ -216,7 +214,7 @@ class GenomicFigure:
             use_human_readable_labels: bool = False,
             tick_height: float = 0.15,
             axis_linewidth: float = 1.1,
-            tick_color: str = 'black',
+            tick_color: str = '#333333',
             tick_linewidth: float = 0.9,
             chromosome_fontweight: FontWeight = FontWeight.BOLD,
             plot_title: bool = True,
@@ -321,7 +319,6 @@ class GenomicFigure:
             font_size: int = 8,
             rect_linewidth: float = 0.7,
             draw_edges: bool = True,
-            label_position: BedLabelPosition = BedLabelPosition.ABOVE,
             plot_title: bool = True,
             plot_scale: bool = True,
             label_on_track: bool = False,
@@ -598,7 +595,6 @@ class GenomicFigure:
             font_size: int = 8,
             rect_linewidth: float = 0.7,
             draw_edges: bool = True,
-            label_position: BedLabelPosition = BedLabelPosition.ABOVE,
             color_by: NarrowPeakColorBy | None = NarrowPeakColorBy.SIGNAL_VALUE,
             cmap: str = 'Oranges',
             min_score: float | None = None,
