+++ plotnado/figure_methods.py
@@ -78,13 +78,11 @@ class GenomicFigureMethods:
         baseline_alpha: float = ...,
         baseline_color: str = ...,
         baseline_linewidth: float = ...,
-        bin_size: int | None = ...,
         color: str = ...,
         fill: bool = ...,
         linewidth: float = ...,
         max_value: float | None = ...,
         min_value: float | None = ...,
-        n_bins: int | None = ...,
         scatter_point_size: float = ...,
         show_baseline: bool = ...,
         smoothing_center: bool = ...,
@@ -381,7 +379,6 @@ class GenomicFigureMethods:
         font_size: int = ...,
         interval_height: float = ...,
         label_field: str = ...,
-        label_position: BedLabelPosition = ...,
         linewidth: float = ...,
         max_rows: int = ...,
         rect_linewidth: float = ...,
@@ -842,7 +839,6 @@ class GenomicFigureMethods:
         font_size: int = ...,
         interval_height: float = ...,
         label_field: str = ...,
-        label_position: BedLabelPosition = ...,
         linewidth: float = ...,
         max_rows: int = ...,
         max_score: float | None = ...,
