+++ plotnado/tracks/bed.py
@@ -13,7 +13,7 @@ from pydantic import ConfigDict, Field, BaseModel
 from .region import GenomicRegion
 from .base import Track, TrackLabeller
 from .utils import clean_axis, read_bed_regions
-from .enums import BedLabelPosition, DisplayMode, TrackType
+from .enums import DisplayMode, TrackType
 from .aesthetics import BaseAesthetics
 from .registry import registry
 
@@ -49,10 +49,6 @@ class BedAesthetics(BaseAesthetics):
     font_size: int = Field(default=8, description="Font size for interval labels.")
     rect_linewidth: float = Field(default=0.7, description="Border line width for interval rectangles.")
     draw_edges: bool = Field(default=True, description="Draw rectangle borders for intervals.")
-    label_position: BedLabelPosition = Field(
-        default=BedLabelPosition.ABOVE,
-        description="Where to place interval labels: 'above', 'inside', or 'below' the rectangle.",
-    )
 
 
 @registry.register(TrackType.BED, aliases=["annotation", "unknown"])
@@ -156,12 +152,6 @@ class BedTrack(Track):
             return
 
         row_scale = 1.0 / max(1, self.max_rows)
-        # In expanded mode cap rect height so rows don't overflow y=[0, 1]
-        effective_height = (
-            self.interval_height
-            if self.display == DisplayMode.COLLAPSED
-            else min(self.interval_height, row_scale * 0.85)
-        )
         row_last_positions: list[int] = []
 
         for row in data.itertuples():
@@ -182,17 +172,11 @@ class BedTrack(Track):
                 else ((row_index + 0.5) * row_scale)
             )
 
-            # Clip to region — intervals may extend beyond the viewed window
-            draw_start = max(float(start), float(gr.start))
-            draw_end = min(float(end), float(gr.end))
-            if draw_end <= draw_start:
-                continue
-
             # Draw interval
             rect = matplotlib.patches.Rectangle(
-                (draw_start, ypos - effective_height / 2),
-                draw_end - draw_start,
-                effective_height,
+                (start, ypos - self.interval_height / 2),
+                end - start,
+                self.interval_height,
                 linewidth=self.rect_linewidth if self.draw_edges else 0,
                 edgecolor=self.edge_color if self.draw_edges else "none",
                 facecolor=self.color,
@@ -202,27 +186,17 @@ class BedTrack(Track):
 
             # Draw label if enabled
             if self.show_labels and hasattr(row, self.label_field):
-                label_text = str(getattr(row, self.label_field))
-                label_xpos = float(start) + float(end)
-                label_xpos /= 2  # true midpoint — allowed to go off-screen
-                match self.label_position:
-                    case BedLabelPosition.ABOVE:
-                        label_ypos = ypos + effective_height / 2 + 0.03
-                        va = "bottom"
-                    case BedLabelPosition.INSIDE:
-                        label_ypos = ypos
-                        va = "center"
-                    case BedLabelPosition.BELOW:
-                        label_ypos = ypos - effective_height / 2 - 0.03
-                        va = "top"
+                label = getattr(row, self.label_field)
+                # Position label above the peak, within track bounds
+                label_ypos = ypos + self.interval_height / 2 + 0.05
                 ax.text(
-                    label_xpos,
+                    (start + end) / 2,
                     label_ypos,
-                    label_text,
+                    str(label),
                     ha="center",
-                    va=va,
+                    va="bottom",
                     fontsize=self.font_size,
-                    clip_on=True,
+                    clip_on=True,  # Clip text that extends outside axis
                 )
 
         ax.set_xlim(gr.start, gr.end)
